1. Modify the serialmotor read to callback;
