+++ include/driver/SerialMotor/Control/Control.cpp
@@ -14,27 +14,27 @@ Control::Control(uint8_t end, uint8_t level) : Serial(end, level){
     
 }
 
-int Control::WritePos(uint8_t ID, uint16_t Position, uint16_t Time, uint16_t Speed)
+int Control::WritePos(uint8_t Id, uint16_t Position, uint16_t Time, uint16_t Speed)
 {
 	uint8_t bBuf[6];
 	Short2Char(bBuf+0, bBuf+1, Position);
 	Short2Char(bBuf+2, bBuf+3, Time);
 	Short2Char(bBuf+4, bBuf+5, Speed);
 	
-	return GenWrite(ID, SERIAL_GOAL_POSITION_L, bBuf, 6);
+	return GenWrite(Id, SERIAL_GOAL_POSITION_L, bBuf, 6);
 }
 
-int Control::RegWritePos(uint8_t ID, uint16_t Position, uint16_t Time, uint16_t Speed)
+int Control::RegWritePos(uint8_t Id, uint16_t Position, uint16_t Time, uint16_t Speed)
 {
 	uint8_t bBuf[6];
 	Short2Char(bBuf+0, bBuf+1, Position);
 	Short2Char(bBuf+2, bBuf+3, Time);
 	Short2Char(bBuf+4, bBuf+5, Speed);
 	
-	return RegWrite(ID, SERIAL_GOAL_POSITION_L, bBuf, 6);
+	return RegWrite(Id, SERIAL_GOAL_POSITION_L, bBuf, 6);
 }
 
-void Control::SyncWritePos(uint8_t ID[], uint8_t IDN, uint16_t Position[], uint16_t Time[], uint16_t Speed[])
+void Control::SyncWritePos(uint8_t Id[], uint8_t IDN, uint16_t Position[], uint16_t Time[], uint16_t Speed[])
 {
     uint8_t offbuf[IDN][6];
     for(uint8_t i = 0; i<IDN; i++){
@@ -55,20 +55,20 @@ void Control::SyncWritePos(uint8_t ID[], uint8_t IDN, uint16_t Position[], uint1
         Short2Char(bBuf+4, bBuf+5, V);
         memcpy(offbuf[i], bBuf, 6);
     }
-    SyncWrite(ID, IDN, SERIAL_GOAL_POSITION_L, (uint8_t*)offbuf, 6);
+    SyncWrite(Id, IDN, SERIAL_GOAL_POSITION_L, (uint8_t*)offbuf, 6);
 }
 
-int Control::PWMMode(uint8_t ID)
+int Control::PWMMode(uint8_t Id)
 {
 	uint8_t bBuf[4];
 	bBuf[0] = 0;
 	bBuf[1] = 0;
 	bBuf[2] = 0;
 	bBuf[3] = 0;
-	return GenWrite(ID, SERIAL_MIN_ANGLE_LIMIT_L, bBuf, 4);	
+	return GenWrite(Id, SERIAL_MIN_ANGLE_LIMIT_L, bBuf, 4);	
 }
 
-int Control::WritePWM(uint8_t ID, int16_t pwmOut)
+int Control::WritePWM(uint8_t Id, int16_t pwmOut)
 {
 	if(pwmOut<0){
 		pwmOut = -pwmOut;
@@ -77,155 +77,87 @@ int Control::WritePWM(uint8_t ID, int16_t pwmOut)
 	uint8_t bBuf[2];
 	Short2Char(bBuf+0, bBuf+1, pwmOut);
 	
-	return GenWrite(ID, SERIAL_GOAL_TIME_L, bBuf, 2);
+	return GenWrite(Id, SERIAL_GOAL_TIME_L, bBuf, 2);
 }
 
-int Control::EnableTorque(uint8_t ID, uint8_t Enable)
+int Control::EnableTorque(uint8_t Id, uint8_t Enable)
 {
-	return WriteByte(ID, SERIAL_TORQUE_ENABLE, Enable);
+	return WriteByte(Id, SERIAL_TORQUE_ENABLE, Enable);
 }
 
-int Control::UnLockEprom(uint8_t ID)
+int Control::UnLockEprom(uint8_t Id)
 {
-	return WriteByte(ID, SERIAL_LOCK, 0);
+	return WriteByte(Id, SERIAL_LOCK, 0);
 }
 
-int Control::LockEprom(uint8_t ID)
+int Control::LockEprom(uint8_t Id)
 {
-	return WriteByte(ID, SERIAL_LOCK, 1);
+	return WriteByte(Id, SERIAL_LOCK, 1);
 }
 
-int Control::FeedBack(int ID)
-{
-	int nLen = Read(ID, SERIAL_PRESENT_POSITION_L, mem, sizeof(mem));
-	if(nLen!=sizeof(mem)){
-		err = 1;
-		return -1;
-	}
-	err = 0;
-	return nLen;
-}
-	
-int Control::ReadPos(int ID)
-{
-	int Pos = -1;
-	if(ID==-1){
-		Pos = mem[SERIAL_PRESENT_POSITION_L-SERIAL_PRESENT_POSITION_L];
-		Pos <<= 8;
-		Pos |= mem[SERIAL_PRESENT_POSITION_H-SERIAL_PRESENT_POSITION_L];
-	}else{
-		err = 0;
-		Pos = ReadWord(ID, SERIAL_PRESENT_POSITION_L);
-		if(Pos==-1){
-			err = 1;
-		}
-	}
-	return Pos;
+int Control::FeedBack(int Id) {
+    int nLen = Read(Id, SERIAL_PRESENT_POSITION_L, mem, sizeof(mem));
+    if (nLen != sizeof(mem)) {
+        err = 1;
+        return -1;
+    }
+
+    ServoState state;
+    state.pos = Char2Short(mem[0], mem[1]);
+    state.speed = Char2Short(mem[2], mem[3]);
+    state.load = Char2Short(mem[4], mem[5]);
+    state.voltage = mem[6];
+    state.temperature = mem[7];
+    state.moving = mem[10];
+    state.current = Char2Short(mem[13], mem[14]);
+
+    {
+        std::lock_guard<std::mutex> lock(servo_mutex);
+        servo_state_map[Id] = state;
+    }
+
+    err = 0;
+    return nLen;
 }
 
-int Control::ReadSpeed(int ID)
-{
-	int Speed = -1;
-	if(ID==-1){
-		Speed = mem[SERIAL_PRESENT_SPEED_L-SERIAL_PRESENT_POSITION_L];
-		Speed <<= 8;
-		Speed |= mem[SERIAL_PRESENT_SPEED_H-SERIAL_PRESENT_POSITION_L];
-	}else{
-		err = 0;
-		Speed = ReadWord(ID, SERIAL_PRESENT_SPEED_L);
-		if(Speed==-1){
-			err = 1;
-			return -1;
-		}
-	}
-	if(!err && (Speed&(1<<15))){
-		Speed = -(Speed&~(1<<15));
-	}	
-	return Speed;
+int Control::ReadPos(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].pos;
+    return -1;
 }
 
-int Control::ReadLoad(int ID)
-{
-	int Load = -1;
-	if(ID==-1){
-		Load = mem[SERIAL_PRESENT_LOAD_L-SERIAL_PRESENT_POSITION_L];
-		Load <<= 8;
-		Load |= mem[SERIAL_PRESENT_LOAD_H-SERIAL_PRESENT_POSITION_L];
-	}else{
-		err = 0;
-		Load = ReadWord(ID, SERIAL_PRESENT_LOAD_L);
-		if(Load==-1){
-			err = 1;
-		}
-	}
-	if(!err && (Load&(1<<10))){
-		Load = -(Load&~(1<<10));
-	}	
-	return Load;
+int Control::ReadSpeed(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].speed;
+    return -1;
 }
 
-int Control::ReadVoltage(int ID)
-{
-	int Voltage = -1;
-	if(ID==-1){
-		Voltage = mem[SERIAL_PRESENT_VOLTAGE-SERIAL_PRESENT_POSITION_L];	
-	}else{
-		err = 0;
-		Voltage = ReadByte(ID, SERIAL_PRESENT_VOLTAGE);
-		if(Voltage==-1){
-			err = 1;
-		}
-	}
-	return Voltage;
+int Control::ReadLoad(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].load;
+    return -1;
 }
 
-int Control::ReadTemper(int ID)
-{
-	int Temper = -1;
-	if(ID==-1){
-		Temper = mem[SERIAL_PRESENT_TEMPERATURE-SERIAL_PRESENT_POSITION_L];	
-	}else{
-		err = 0;
-		Temper = ReadByte(ID, SERIAL_PRESENT_TEMPERATURE);
-		if(Temper==-1){
-			err = 1;
-		}
-	}
-	return Temper;
+int Control::ReadVoltage(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].voltage;
+    return -1;
 }
 
-int Control::ReadMove(int ID)
-{
-	int Move = -1;
-	if(ID==-1){
-		Move = mem[SERIAL_MOVING-SERIAL_PRESENT_POSITION_L];	
-	}else{
-		err = 0;
-		Move = ReadByte(ID, SERIAL_MOVING);
-		if(Move==-1){
-			err = 1;
-		}
-	}
-	return Move;
+int Control::ReadTemper(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].temperature;
+    return -1;
 }
 
-int Control::ReadCurrent(int ID)
-{
-	int Current = -1;
-	if(ID==-1){
-		Current = mem[SERIAL_PRESENT_CURRENT_L-SERIAL_PRESENT_POSITION_L];
-		Current <<= 8;
-		Current |= mem[SERIAL_PRESENT_CURRENT_H-SERIAL_PRESENT_POSITION_L];
-	}else{
-		err = 0;
-		Current = ReadWord(ID, SERIAL_PRESENT_CURRENT_L);
-		if(Current==-1){
-			err = 1;
-			return -1;
-		}
-	}
-	if(!err && (Current&(1<<15))){
-		Current = -(Current&~(1<<15));
-	}	
-	return Current;
+int Control::ReadMove(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].moving;
+    return -1;
+}
+
+int Control::ReadCurrent(int Id) {
+    std::lock_guard<std::mutex> lock(servo_mutex);
+    if (servo_state_map.count(Id)) return servo_state_map[Id].current;
+    return -1;
 }
